I got the camera and odomerry tracking wokring together, i ended up messing up the merging eqation and then also had to do some typos with bpthj csmera and odomertyr tracking with soem wrong eqations yk

+++ src/mainFunctionsC/functions.cpp
@@ -31,9 +31,11 @@ double functions::normalizeDegAngle(double A){
 //quick rounding to the nearist thousond.
 double functions::roundNearistThous(double Num){
     double n = Num;
-    n = round(n*1000);
+    n = (double)round(n*1000);
     n = n/1000;
     return n;
+
+    
 }
 
 // milameters to inches
